i have trained emnist with convolution layer

+++ main.hpp
@@ -1,5 +1,5 @@
 #pragma once
 // using valT = double;
-// using valT = long double;
-using valT = float;
+using valT = long double;
+// using valT = float;
 using funcT = valT (*)(valT);
